refactor(orchestration): extract RpcDelegationService
Eight methods, 570 body lines, moved verbatim, no receiver rewrite needed.
1399 -> 827 lines. One public method; the other seven are private, called only
from within the eight, and stay private with no facade delegation.

The plan's port list named ensureWorkerSession, cancelWorkerTask,
closeWorkerSession and logger. None of the eight uses any of them --
ensureWorkerSession is reached through workerSessions.ensureReservedWorkerSession,
and every log goes through kernel.logEvent. Six ports.

requestDelegateFromRpc ends with `void this.runAutoRunRpcWorkerTask({...})`,
detached and outside every mutate. runAutoRunRpcWorkerTask opens its own, so a
chain created inside a critical section would inherit the AsyncLocalStorage
store and throw. It moved untouched.

The requestDelegate overload dispatcher now calls this.rpcDelegation directly
rather than the facade delegation, matching how every other service reaches its
collaborators. I expected this to be load-bearing -- Task 14 saw an added
microtask hop reorder a detached chain against a synchronous log -- but it is
not: routing the dispatcher through the delegation leaves all 24 fixtures and
all 185 oracle tests green. This path does not observe hop depth. cancelGroup
remains the only one that does.
